requirements.txt
# Claude AI
anthropic>=0.25.0

# PDF text extraction
pypdf2>=3.0.0
pdfplumber>=0.10.0

# OCR/ICR support
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0

# File watching
watchdog>=4.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
